requirements.txt
# Core dependencies
gradio>=5.13.0
gradio-client>=1.6.0
fastapi>=0.68.0
python-dotenv>=0.19.0
uvicorn>=0.15.0

# AI/ML dependencies
google-generativeai>=0.3.0  # For Gemini
anthropic>=0.3.0  # For Claude
e2b>=0.9.0  # For secure sandbox

# Utility dependencies
python-logging>=0.4.9
aiohttp>=3.8.0
requests>=2.26.0
typing-extensions>=4.0.0
pydantic>=1.8.0

# Development dependencies
pytest>=6.2.5
pytest-asyncio>=0.16.0
black>=21.9b0
flake8>=3.9.2
mypy>=0.910 